@@ -1,6 +1,6 @@
 {
   "name": "@jridgewell/trace-mapping",
-  "version": "0.3.14",
+  "version": "0.3.18",
   "description": "Trace the original position through a source map",
   "keywords": [
     "source",
@@ -8,10 +8,9 @@
   ],
   "main": "dist/trace-mapping.umd.js",
   "module": "dist/trace-mapping.mjs",
-  "typings": "dist/types/trace-mapping.d.ts",
+  "types": "dist/types/trace-mapping.d.ts",
   "files": [
-    "dist",
-    "src"
+    "dist"
   ],
   "exports": {
     ".": [
@@ -53,24 +52,24 @@
     "test:watch": "ava --watch"
   },
   "devDependencies": {
-    "@rollup/plugin-typescript": "8.3.2",
-    "@typescript-eslint/eslint-plugin": "5.23.0",
-    "@typescript-eslint/parser": "5.23.0",
-    "ava": "4.2.0",
+    "@rollup/plugin-typescript": "8.5.0",
+    "@typescript-eslint/eslint-plugin": "5.39.0",
+    "@typescript-eslint/parser": "5.39.0",
+    "ava": "4.3.3",
     "benchmark": "2.1.4",
-    "c8": "7.11.2",
-    "esbuild": "0.14.38",
-    "esbuild-node-loader": "0.8.0",
-    "eslint": "8.15.0",
+    "c8": "7.12.0",
+    "esbuild": "0.15.10",
+    "eslint": "8.25.0",
     "eslint-config-prettier": "8.5.0",
-    "eslint-plugin-no-only-tests": "2.6.0",
+    "eslint-plugin-no-only-tests": "3.0.0",
     "npm-run-all": "4.1.5",
-    "prettier": "2.6.2",
-    "rollup": "2.72.1",
-    "typescript": "4.6.4"
+    "prettier": "2.7.1",
+    "rollup": "2.79.1",
+    "tsx": "3.10.1",
+    "typescript": "4.8.4"
   },
   "dependencies": {
-    "@jridgewell/resolve-uri": "^3.0.3",
-    "@jridgewell/sourcemap-codec": "^1.4.10"
+    "@jridgewell/resolve-uri": "3.1.0",
+    "@jridgewell/sourcemap-codec": "1.4.14"
   }
 }
