@@ -1,40 +1,44 @@
-import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
+import {BrowserModule} from '@angular/platform-browser';
+import {NgModule} from '@angular/core';
 
-import { AppComponent } from './app.component';
-import { ProductListComponent } from './components/product-list/product-list.component';
+import {AppComponent} from './app.component';
+import {ProductListComponent} from './components/product-list/product-list.component';
 import {ProductCategoryMenuComponent} from "./components/product-category-menu/product-category-menu.component";
-import { HttpClientModule } from '@angular/common/http';
-import { ProductService } from './services/product.service';
+import {HttpClientModule} from '@angular/common/http';
+import {ProductService} from './services/product.service';
 
-import { Routes, RouterModule} from '@angular/router';
-import { SearchComponent } from './components/search/search.component';
-import { ProductDetailsComponent } from './components/product-details/product-details.component';
+import {Routes, RouterModule} from '@angular/router';
+import {SearchComponent} from './components/search/search.component';
+import {ProductDetailsComponent} from './components/product-details/product-details.component';
+
+import {NgbModule} from "@ng-bootstrap/ng-bootstrap";
 
 const routes: Routes = [
-  {path: 'products/:id', component:ProductDetailsComponent},
-  {path: 'search/:keyword', component:ProductListComponent},
-  {path: 'category/:id/:name', component: ProductListComponent},
-  {path: 'category', component: ProductListComponent},
-  {path: 'products', component: ProductListComponent},
-  {path: '', redirectTo: '/products', pathMatch: 'full'},
-  {path: '**', redirectTo: '/products', pathMatch: 'full'}
+    {path: 'products/:id', component: ProductDetailsComponent},
+    {path: 'search/:keyword', component: ProductListComponent},
+    {path: 'category/:id/:name', component: ProductListComponent},
+    {path: 'category', component: ProductListComponent},
+    {path: 'products', component: ProductListComponent},
+    {path: '', redirectTo: '/products', pathMatch: 'full'},
+    {path: '**', redirectTo: '/products', pathMatch: 'full'}
 ];
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    ProductListComponent,
-    ProductCategoryMenuComponent,
-    SearchComponent,
-    ProductDetailsComponent
-  ],
-  imports: [
-    RouterModule.forRoot(routes),
-    BrowserModule,
-    HttpClientModule
-  ],
-  providers: [ProductService],
-  bootstrap: [AppComponent]
+    declarations: [
+        AppComponent,
+        ProductListComponent,
+        ProductCategoryMenuComponent,
+        SearchComponent,
+        ProductDetailsComponent
+    ],
+    imports: [
+        RouterModule.forRoot(routes),
+        BrowserModule,
+        HttpClientModule,
+        NgbModule
+    ],
+    providers: [ProductService],
+    bootstrap: [AppComponent]
 })
-export class AppModule { }
+export class AppModule {
+}
