@@ -1,65 +1,94 @@
-import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { Product } from '../common/product';
-import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
-import { ProductCategory } from "../../common/product-category";
+import {Injectable} from '@angular/core';
+import {HttpClient} from '@angular/common/http';
+import {Product} from '../common/product';
+import {Observable} from 'rxjs';
+import {map} from 'rxjs/operators';
+import {ProductCategory} from "../../common/product-category";
 
 @Injectable({
-  providedIn: 'root'
+    providedIn: 'root'
 })
 export class ProductService {
 
-  private baseUrl = 'http://localhost:8080/api/products';
+    private baseUrl = 'http://localhost:8080/api/products';
 
-  private categoryUrl = 'http://localhost:8080/api/product-category';
+    private categoryUrl = 'http://localhost:8080/api/product-category';
 
-  constructor(private httpClient: HttpClient) { }
+    constructor(private httpClient: HttpClient) {
+    }
 
-  getProductList(theCategoryId: number): Observable<Product[]> {
+    getProductListPaginate(thePage: number,
+                           thePageSize: number,
+                           theCategoryId: number): Observable<GetResponseProducts> {
 
-    // need to build URL based on category id
-    const searchUrl = `${this.baseUrl}/search/findByCategoryId?id=${theCategoryId}`;
+        // need to build URL based on category id,page  and page size number
+        const searchUrl = `${this.baseUrl}/search/findByCategoryId?id=${theCategoryId}`
+            + `&page=${thePage}&size=${thePageSize}`;
 
-    return this.getProducts(searchUrl);
-  }
+        return this.httpClient.get<GetResponseProducts>(searchUrl);
+    }
 
-  searchProducts(theKeyword: string): Observable<Product[]> {
+    getProductList(theCategoryId: number): Observable<Product[]> {
 
-    // need to build URL based on the keyword
-    const searchUrl = `${this.baseUrl}/search/findByNameContaining?name=${theKeyword}`;
+        // need to build URL based on category id
+        const searchUrl = `${this.baseUrl}/search/findByCategoryId?id=${theCategoryId}`;
 
-    return this.getProducts(searchUrl);
-  }
+        return this.getProducts(searchUrl);
+    }
 
-  private getProducts(searchUrl: string): Observable<Product[]> {
-    return this.httpClient.get<GetResponseProducts>(searchUrl).pipe(map(response => response._embedded.products));
-  }
+    searchProducts(theKeyword: string): Observable<Product[]> {
 
-  getProductCategories(): Observable<ProductCategory[]> {
+        // need to build URL based on the keyword
+        const searchUrl = `${this.baseUrl}/search/findByNameContaining?name=${theKeyword}`;
 
-    return this.httpClient.get<GetResponseProductCategory>(this.categoryUrl).pipe(
-        map(response => response._embedded.productCategory)
-    );
-  }
+        return this.getProducts(searchUrl);
+    }
 
-    getProduct(theProductId: number):Observable<Product> {
-    //The template literal ${this.baseUrl}/${theProductId}
-      // combines the value of this.baseUrl and theProductId to form the URL for the GET request.
-    const productUrl = `${this.baseUrl}/${theProductId}`;
+    searchProductsPaginate(thePage: number,
+                           thePageSize: number,
+                           theKeyword:string): Observable<GetResponseProducts> {
 
-    return this.httpClient.get<Product>(productUrl);
+        // need to build URL based on keyword,page  and page size number
+        const searchUrl = `${this.baseUrl}/search/findByNameContaining?name=${theKeyword}`
+            + `&page=${thePage}&size=${thePageSize}`;
+
+        return this.httpClient.get<GetResponseProducts>(searchUrl);
+    }
+
+    private getProducts(searchUrl: string): Observable<Product[]> {
+        return this.httpClient.get<GetResponseProducts>(searchUrl).pipe(map(response => response._embedded.products));
+    }
+
+    getProductCategories(): Observable<ProductCategory[]> {
+
+        return this.httpClient.get<GetResponseProductCategory>(this.categoryUrl).pipe(
+            map(response => response._embedded.productCategory)
+        );
+    }
+
+    getProduct(theProductId: number): Observable<Product> {
+        //The template literal ${this.baseUrl}/${theProductId}
+        // combines the value of this.baseUrl and theProductId to form the URL for the GET request.
+        const productUrl = `${this.baseUrl}/${theProductId}`;
+
+        return this.httpClient.get<Product>(productUrl);
     }
 }
 
 interface GetResponseProducts {
-  _embedded: {
-    products: Product[];
-  }
+    _embedded: {
+        products: Product[];
+    },
+    page: {
+        size: number,
+        totalElements: number,
+        totalPages: number,
+        number: number
+    }
 }
 
 interface GetResponseProductCategory {
-  _embedded: {
-    productCategory: ProductCategory[];
-  }
+    _embedded: {
+        productCategory: ProductCategory[];
+    }
 }
