@@ -1,29 +1,31 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 import {Product} from "../../common/product";
 import {ProductService} from "../../services/product.service";
 import {ActivatedRoute} from "@angular/router";
 
 @Component({
-  selector: 'app-product-details',
-  templateUrl: './product-details.component.html',
-  styleUrls: ['./product-details.component.css']
+    selector: 'app-product-details',
+    templateUrl: './product-details.component.html',
+    styleUrls: ['./product-details.component.css']
 })
 export class ProductDetailsComponent implements OnInit {
 
-  product!:Product;
-  constructor(private productService:ProductService,
-              private route:ActivatedRoute) { }
+    product!: Product;
 
-  ngOnInit(): void {
-    this.route.paramMap.subscribe(()=>{
-      this.handleProductDetails();
-    })
-  }
+    constructor(private productService: ProductService,
+                private route: ActivatedRoute) {
+    }
 
-  private handleProductDetails() {
-    //get the "id" param string and convert it to a number
-    const theProductId:number = +this.route.snapshot.paramMap.get('id')!;
-    this.productService.getProduct(theProductId).subscribe(data=>
-    this.product = data)
-  }
+    ngOnInit(): void {
+        this.route.paramMap.subscribe(() => {
+            this.handleProductDetails();
+        })
+    }
+
+    private handleProductDetails() {
+        //get the "id" param string and convert it to a number
+        const theProductId: number = +this.route.snapshot.paramMap.get('id')!;
+        this.productService.getProduct(theProductId).subscribe(data =>
+            this.product = data)
+    }
 }
